refactor: Add AddDisabilities component for adding disabilities
The code changes introduce a new component called `AddDisabilities.vue` that allows users to add disabilities. This component includes a dialog panel with input fields for entering the name of the disability. The `createDisabilitie` function handles the creation of the disability by dispatching the `CREATE_DISABILITY` action. The component also emits a `close-add` event when the dialog is closed.

Closes #789